+++ create_docs.sh
@@ -5,5 +5,6 @@ pydoc-markdown \
   -m tankobon.iso639 \
   -m tankobon.models \
   -m tankobon.utils \
+  -m tankobon.imposter \
   -m tankobon.sources.base \
   --render-toc > API.md