+++ create_resources.py
@@ -16,7 +16,7 @@ QRC_TEMPLATE = """
 CURDIR = pathlib.Path(__file__).parent
 COMPILER = "pyside6-rcc"
 RESOURCES = CURDIR / "resources"
-DESTINATION = CURDIR / "tankobon" / "gui" / "resources.py"
+DESTINATION = CURDIR / "tankobon" / "ui" / "resources.py"
 
 
 def create_qrc():