+++ tankobon/sources/__init__.py
@@ -32,4 +32,4 @@ manga = parser.create(manga_url)
 """
 
 # import to register default sources
-from . import catmanga, genshin, mangadex, mangakakalot
+from . import mangadex, mangakakalot