+++ tests/test_imposter.py
@@ -1,0 +1,8 @@
+# coding: utf8
+
+from tankobon import imposter
+
+
+def test_random():
+    ua = imposter.UserAgent()
+    print(ua.random())